+++ cfgov/cfgov/settings/test.py
@@ -7,4 +7,8 @@ DATABASES = {
     }
 }
 
+EMAIL_BACKEND = 'django.core.mail.backends.locmem.EmailBackend'
+
 TEST_RUNNER = 'cfgov.test.TestDataTestRunner'
+
+LOGGING = {}