+++ cfgov/v1/atomic_elements/molecules.py
@@ -13,12 +13,22 @@ def isRequired(field_name):
 
 class HalfWidthLinkBlob(blocks.StructBlock):
     heading = blocks.CharBlock(required=False)
+    sub_heading = blocks.CharBlock(required=False)
+    sub_heading_icon = blocks.CharBlock(
+        required=False,
+        label="Sub heading icon",
+        help_text=(
+            'A list of icon names can be obtained at: '
+            'https://cfpb.github.io/capital-framework/components/cf-icons/. '
+            'Examples: linkedin-square, facebook-square, etc.'
+            )
+        )
     body = blocks.RichTextBlock(blank=True, required=False)
     links = blocks.ListBlock(atoms.Hyperlink(), required=False)
 
     class Meta:
         icon = 'link'
-        template = '_includes/molecules/half-width-link-blob.html'
+        template = '_includes/molecules/link-blob.html'
 
 
 class ImageText5050(blocks.StructBlock):
@@ -139,7 +149,7 @@ class CallToAction(blocks.StructBlock):
     class Meta:
         template = '_includes/molecules/call-to-action.html'
         icon = 'grip'
-        label = 'Call to Action'
+        label = 'Call to action'
 
 
 class ContactAddress(blocks.StructBlock):
@@ -187,8 +197,9 @@ class RelatedLinks(blocks.StructBlock):
     links = blocks.ListBlock(atoms.Hyperlink())
 
     class Meta:
-        icon = 'link'
-        template = '_includes/molecules/related-links.html'
+        icon = 'grip'
+        template = '_includes/molecules/related-content.html'
+        label = 'Related content'
 
 
 class Quote(blocks.StructBlock):
@@ -225,7 +236,8 @@ class RelatedMetadata(blocks.StructBlock):
     class Meta:
         icon = 'grip'
         template = '_includes/molecules/related-metadata.html'
-        label = 'Related Metadata'
+        label = 'Related metadata'
+
 
 class RSSFeed(blocks.ChoiceBlock):
     choices = [
@@ -236,6 +248,7 @@ class RSSFeed(blocks.ChoiceBlock):
     class Meta:
         icon = 'plus'
         template = '_includes/molecules/rss-feed.html'
+        label = 'RSS feed'
 
 
 class SocialMedia(blocks.StructBlock):