+++ cfgov/v1/atomic_elements/organisms.py
@@ -1,3 +1,6 @@
+from django.apps import apps
+from django.utils.encoding import smart_text
+from wagtail.contrib.table_block.blocks import TableBlock
 from wagtail.wagtailcore import blocks
 from wagtail.wagtailimages import blocks as images_blocks
 from wagtail.wagtailsnippets.blocks import SnippetChooserBlock
@@ -18,6 +21,19 @@ class Well(blocks.StructBlock):
 
 class ImageText5050Group(blocks.StructBlock):
     heading = blocks.CharBlock(icon='title', required=False)
+
+    sharing = blocks.StructBlock([
+        ('shareable', blocks.BooleanBlock(label='Include sharing links?',
+                                          help_text='If checked, share links '
+                                                    'will be included below '
+                                                    'the items.',
+                                          required=False)),
+        ('share_blurb', blocks.CharBlock(help_text='Sets the tweet text, '
+                                                   'email subject line, and '
+                                                   'LinkedIn post text.',
+                                         required=False)),
+    ])
+
     image_texts = blocks.ListBlock(molecules.ImageText5050())
 
     class Meta:
@@ -34,10 +50,20 @@ class ImageText2575Group(blocks.StructBlock):
         template = '_includes/organisms/image-text-25-75-group.html'
 
 
-class HalfWidthLinkBlobGroup(blocks.StructBlock):
+class LinkBlobGroup(blocks.StructBlock):
     heading = blocks.CharBlock(icon='title', required=False)
+    has_top_border = blocks.BooleanBlock(required=False)
+    has_bottom_border = blocks.BooleanBlock(required=False)
     link_blobs = blocks.ListBlock(molecules.HalfWidthLinkBlob())
 
+
+class ThirdWidthLinkBlobGroup(LinkBlobGroup):
+    class Meta:
+        icon = 'link'
+        template = '_includes/organisms/third-width-link-blob-group.html'
+
+
+class HalfWidthLinkBlobGroup(LinkBlobGroup):
     class Meta:
         icon = 'link'
         template = '_includes/organisms/half-width-link-blob-group.html'
@@ -129,6 +155,155 @@ class Table(blocks.StructBlock):
         icon = 'form'
         template = '_includes/organisms/table.html'
         label = 'Table'
+
+
+class ModelBlock(blocks.StructBlock):
+    """Abstract StructBlock that provides Django model instances to subclasses.
+
+    This class inherits from the standard Wagtail StructBlock but adds helper
+    methods that allow subclasses to dynamically render Django model instances.
+    This is useful if, for example, a widget needs to show a list of all model
+    instances meeting a certain criteria.
+
+    Subclasses must override the 'model' class attribute with the fully-
+    qualified name of the model to be used, for example 'my.app.Modelname'.
+
+    Subclasses may optionally override the 'filter_queryset' method to do
+    filtering on the model QuerySet.
+
+    Subclasses may optionally override either the class attributes 'ordering'
+    (providing a Django-style string or tuple of orderings to use) and 'limit'
+    (providing an integer to use to slice the model QuerySet), or provide
+    methods 'get_ordering' and 'get_limit' that do the same thing.
+
+    """
+    model = None
+    ordering = None
+    limit = None
+
+    def get_queryset(self, value):
+        model_cls = apps.get_model(self.model)
+        qs = model_cls.objects.all()
+
+        qs = self.filter_queryset(qs, value)
+
+        ordering = self.get_ordering(value)
+        if ordering:
+            if isinstance(ordering, basestring):
+                ordering = (ordering,)
+
+            qs = qs.order_by(*ordering)
+
+        limit = self.get_limit(value)
+        if limit:
+            qs = qs[:limit]
+
+        return qs
+
+    def filter_queryset(self, qs, value):
+        return qs
+
+    def get_ordering(self, value):
+        return self.ordering
+
+    def get_limit(self, value):
+        return self.limit
+
+
+class ModelTable(ModelBlock):
+    """Abstract StructBlock that can generate a table from a Django model.
+
+    Subclasses must override the 'fields' and 'field_headers' class attributes
+    to specify which model fields to include in the generated table.
+
+    By default model instance values will be converted to text for display
+    in table rows. To override this, subclasses may define custom field
+    formatter methods, using the name 'make_FIELD_value'. This may be useful
+    if fields are non-text types, for example when formatting dates.
+
+    For example:
+
+        def make_created_value(self, instance, value):
+            return value.strftime('%b %d, %Y')
+
+    """
+    fields = None
+    field_headers = None
+
+    row_links = blocks.BooleanBlock(
+        required=False,
+        default=True,
+        help_text='Whether to highlight rows containing links'
+    )
+
+    def render(self, value):
+        rows = [
+            self.make_row(instance)
+            for instance in self.get_queryset(value)
+        ]
+
+        table_value = {
+            'headers': self.field_headers,
+            'rows': rows,
+            'row_links': value.get('row_links'),
+        }
+
+        table = Table()
+        value = table.to_python(table_value)
+        return table.render(value)
+
+    def make_row(self, instance):
+        return [
+            {'type': 'text', 'value': self.make_value(instance, field)}
+            for field in self.fields
+        ]
+
+    def make_value(self, instance, field):
+        value = getattr(instance, field)
+        return self.format_field_value(instance, field, value)
+
+    def format_field_value(self, instance, field, value):
+        custom_formatter_name = 'make_{}_value'.format(field)
+        custom_formatter = getattr(self, custom_formatter_name, None)
+
+        if custom_formatter:
+            return custom_formatter(instance, value)
+        else:
+            return smart_text(value)
+
+    class Meta:
+        icon = 'table'
+
+
+class ModelList(ModelBlock):
+    """Abstract StructBlock that can generate a list from a Django model.
+
+    Subclasses must define a 'render' method that renders the model QuerySet
+    to a string.
+
+    For example:
+
+        def render(self, value):
+            value['objects'] = self.get_queryset(value)
+            template = 'path/to/template.html'
+            return render_to_string(template, value)
+
+    """
+    limit = atoms.IntegerBlock(
+        default=5,
+        label='Maximum items',
+        min_value=0,
+        help_text='Limit list to this number of items'
+    )
+
+    def __init__(self, *args, **kwargs):
+        super(ModelList, self).__init__(*args, **kwargs)
+
+    def get_limit(self, value):
+        return value.get('limit')
+
+    class Meta:
+        icon = 'list-ul'
 
 
 class FullWidthText(blocks.StreamBlock):