+++ cfgov/v1/auth_forms.py
@@ -1,19 +1,17 @@
 import time
+
 from datetime import timedelta
-
-from django.contrib.auth.forms import (PasswordChangeForm, PasswordResetForm,
-                                       SetPasswordForm)
-
-from django.contrib.auth import authenticate
-from django.contrib.auth import get_user_model
-from django.utils import timezone
 from django.conf import settings
+from django.contrib.auth import authenticate, get_user_model
+from django.contrib.auth.forms import (
+    AuthenticationForm, PasswordChangeForm, PasswordResetForm,
+    SetPasswordForm
+)
 from django.core.exceptions import ValidationError, ObjectDoesNotExist
-from django.contrib.auth.forms import AuthenticationForm
+from django.utils import timezone
+from wagtail.wagtailusers import forms as wagtailforms
 
-from wagtail.wagtailadmin import forms as wagtail_adminforms
-from wagtail.wagtailusers.forms import UserCreationForm, UserEditForm
-
+from .email import send_password_reset_email
 from .models import base
 from .util import password_policy
 
@@ -122,4 +120,35 @@ class LoginForm(AuthenticationForm):
             raise ValidationError("This account is temporarily locked; please try later or <a href='/admin/password_reset/' style='color:white;font-weight:bold'>reset your password</a>")
 
 
+class UserCreationForm(wagtailforms.UserCreationForm):
+    def clean_email(self):
+        email = self.cleaned_data['email']
 
+        try:
+            User = get_user_model()
+            User.objects.get(email=email)
+        except User.DoesNotExist:
+            return email
+        else:
+            raise ValidationError('This email is already in use.')
+
+    def save(self, commit=True):
+        user = super(UserCreationForm, self).save(commit=commit)
+
+        if commit:
+            send_password_reset_email(user.email)
+
+        return user
+
+
+class UserEditForm(wagtailforms.UserEditForm):
+    def clean_email(self):
+        email = self.cleaned_data['email']
+
+        try:
+            User = get_user_model()
+            User.objects.exclude(pk=self.instance.pk).get(email=email)
+        except User.DoesNotExist:
+            return email
+        else:
+            raise ValidationError('This email is already in use.')