+++ cfgov/v1/models/sublanding_page.py
@@ -1,5 +1,3 @@
-from itertools import chain
-
 from wagtail.wagtailcore.fields import StreamField
 from wagtail.wagtailadmin.edit_handlers import TabbedInterface, ObjectList, \
     StreamFieldPanel
@@ -9,6 +7,7 @@ from wagtail.wagtailimages.blocks import ImageChooserBlock
 from .base import CFGOVPage
 from ..atomic_elements import molecules, organisms
 from ..util import filterable_list, util
+from jobmanager.models import JobListingList
 
 class SublandingPage(CFGOVPage):
     header = StreamField([
@@ -21,6 +20,7 @@ class SublandingPage(CFGOVPage):
         ('image_text_50_50_group', organisms.ImageText5050Group()),
         ('full_width_text', organisms.FullWidthText()),
         ('half_width_link_blob_group', organisms.HalfWidthLinkBlobGroup()),
+        ('third_width_link_blob_group', organisms.ThirdWidthLinkBlobGroup()),
         ('post_preview_snapshot', organisms.PostPreviewSnapshot()),
         ('well', organisms.Well()),
         ('table', organisms.Table()),
@@ -41,6 +41,7 @@ class SublandingPage(CFGOVPage):
             ('body', blocks.TextBlock(required=False, label='Introduction Body')),
         ], heading='Breakout Image', icon='image')),
         ('related_posts', organisms.RelatedPosts()),
+        ('job_listing_list', JobListingList()),
     ], blank=True)
 
     # General content tab
@@ -64,18 +65,15 @@ class SublandingPage(CFGOVPage):
 
     def get_browsefilterable_posts(self, request, limit):
         filter_pages = [p.specific for p in self.get_appropriate_descendants(request.site.hostname)
-                        if 'FilterablePage' in p.specific_class.__name__ and 'archive' not in p.title.lower()]
-        filtered_controls = {}
-        for page in filter_pages:
-            id = str(util.get_form_id(page))
-            if id not in filtered_controls.keys():
-                filtered_controls.update({id: []})
-            form_class = page.get_form_class()
-            posts = page.get_page_set(form_class(parent=page, hostname=request.site.hostname), request.site.hostname)
-            if filtered_controls[id]:
-                filtered_controls[id] += posts
-            else:
-                filtered_controls[id] = posts
-        posts_tuple_list = [(id, post) for id, posts in filtered_controls.iteritems() for post in posts]
+                        if 'FilterablePage' in p.specific_class.__name__
+                        and 'archive' not in p.title.lower()]
+
+        posts_tuple_list = [(str(util.get_form_id(page)), post)
+                            for page in filter_pages
+                            for post in page.get_page_set(
+                                page.get_form_class()(parent=page,
+                                                      hostname=request.site.hostname),
+                                request.site.hostname)]
+
         posts = sorted(posts_tuple_list, key=lambda p: p[1].date_published, reverse=True)[:limit]
         return posts