+++ gulp/tasks/styles.js
@@ -3,6 +3,7 @@
 var gulp = require( 'gulp' );
 var plugins = require( 'gulp-load-plugins' )();
 var mqr = require( 'gulp-mq-remove' );
+var cleanCSS = require('gulp-clean-css');
 var config = require( '../config' );
 var configPkg = config.pkg;
 var configBanner = config.banner;
@@ -50,7 +51,7 @@ function stylesIe() {
       width: '75em'
     } ) )
     // mqr expands the minified file
-    .pipe( plugins.cssmin() )
+    .pipe( cleanCSS( {compatibility: 'ie8'} ) )
     .pipe( plugins.rename( {
       suffix:  '.ie',
       extname: '.css'
@@ -82,12 +83,33 @@ function stylesOnDemand() {
       width: '75em'
     } ) )
     // mqr expands the minified file
-    .pipe( plugins.cssmin() )
+    .pipe( cleanCSS( {compatibility: 'ie8'} ) )
     .pipe( plugins.rename( {
       suffix:  '.nonresponsive',
       extname: '.css'
     } ) )
     .pipe( gulp.dest( configStyles.dest ) )
+    .pipe( browserSync.reload( {
+      stream: true
+    } ) );
+}
+
+/**
+ * Process CSS for Wagtail feature flags.
+ * @returns {PassThrough} A source stream.
+ */
+function stylesFeatureFlags() {
+  return gulp.src( configStyles.cwd + '/feature-flags/*.less' )
+    .pipe( plugins.less( configStyles.settings ) )
+    .on( 'error', handleErrors )
+    .pipe( plugins.autoprefixer( {
+      browsers: [ 'last 2 version',
+                  'ie 7-8',
+                  'android 4',
+                  'BlackBerry 7',
+                  'BlackBerry 10' ]
+    } ) )
+    .pipe( gulp.dest( configStyles.dest + '/feature-flags' ) )
     .pipe( browserSync.reload( {
       stream: true
     } ) );
@@ -142,6 +164,7 @@ function stylesNemoIE() {
 gulp.task( 'styles:modern', stylesModern );
 gulp.task( 'styles:ie', stylesIe );
 gulp.task( 'styles:ondemand', stylesOnDemand );
+gulp.task( 'styles:featureFlags', stylesFeatureFlags );
 gulp.task( 'styles:nemoProd', stylesNemoProd );
 gulp.task( 'styles:nemoIE', stylesNemoIE );
 gulp.task( 'styles:nemo', [
@@ -153,5 +176,6 @@ gulp.task( 'styles', [
   'styles:modern',
   'styles:ie',
   'styles:ondemand',
+  'styles:featureFlags',
   'styles:nemo'
 ] );